Update for V1.3.5
The big change here was making a separate pitch for the threads & the helix.  That way a bigger pitch for the helix does not effect the threads.  A larger helix pitch than the thread pitch will allow for the ability to make the M/F Thread offset the same distance all the way around the thread.